% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1a_cleanData.R
\name{addGreppedPatterns}
\alias{addGreppedPatterns}
\title{Search for inventory-related information within item titles}
\usage{
addGreppedPatterns(itemsTuple)
}
\arguments{
\item{itemsTuple}{name of dataframe with items data, described above}
}
\value{
dataframe with six columns added
}
\description{
Given items data, with columns `hash_str`, `title`, `prediction`, extract
columns grams, ounces, pounds, milligrams, pills, quantity
}
\details{
Grams is the information associated with grams, e.g. "half gram" or "10g",
etc., and similarly for ounces and pounds. Milligrams can be both milligrams
and micrograms. Pills refers to pills, tabs, tablet, blotters. In the absence
of any of the preceding information, we extract a beginning numeric quantity
in the item title. For example, in the title "10 credit cards," 10 will be
extracted.
}
