% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1a_cleanData.R
\name{cleanFeedback}
\alias{cleanFeedback}
\title{Clean feedback data}
\usage{
cleanFeedback(feedback, items)
}
\arguments{
\item{feedback}{name of dataframe with feedback data as described above}

\item{items}{name of dataframe with unique items, needs to contain
`item_hash` and `vendor_hash` at a minimum}
}
\value{
dataframe with usable feedback data
}
\description{
Given feedback data, with columns `hash_str`, `marketplace`, `item_hash`,
`date`, `order_amount_usd`, add `vendor_hash` to link to users data, and delete
rows that don't have a vendor match
}
