% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1a_cleanData.R
\name{cleanItems}
\alias{cleanItems}
\title{Clean items data}
\usage{
cleanItems(items, feedback)
}
\arguments{
\item{items}{name of dataframe with unique items, described above}

\item{feedback}{name of dataframe with feedback data, needs at least
`item_hash`}
}
\value{
dataframe with usable items data
}
\description{
Given items data, with columns `hash_str`, `marketplace`, `title`, `vendor`,
`vendor_hash`, `prediction`, extract inventory, i.e. add `dosage` and `unit`.
Delete rows that have no feedback, i.e. no sales, since we will not be
analyzing these.
}
