% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1b_cleanParsedData.R
\name{cleanParsedItems}
\alias{cleanParsedItems}
\title{Clean parsed item scrapes}
\usage{
cleanParsedItems(parsedItems, items)
}
\arguments{
\item{parsedItems}{name of dataframe that contains parsed item listing
scrapes, as described above.}

\item{items}{name of dataframe with unique item listings, with at least
columns `hash_str`, `marketplace`, `title`, `vendor`}
}
\value{
1. `parsedUsers` dataframe with `listing_description` column replaced
  by `descriptionClean`, which has any PGP keys extracted. 2. `retrievedPGPs`
  dataframe with columns `date`, `vendor_hash` and `PGPclean` (extracted PGP
  keys)
}
\description{
Given a parsed item scrapes, containing columns `marketplace`, `date`,
`seller_id`, `title`, `listing_description`, remove rows with missing
information, or duplicated rows, match to items database and remove
non-matches, and extract any available PGP keys. This is analogous to
`cleanParsedUsers()`.
}
