% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1b_cleanParsedData.R
\name{cleanParsedUsers}
\alias{cleanParsedUsers}
\title{Clean parsed profile scrapes}
\usage{
cleanParsedUsers(parsedUsers, users)
}
\arguments{
\item{parsedUsers}{name of dataframe that contains parsed profile scrapes, as
described above.}

\item{users}{name of dataframe with unique user accounts, with columns
`hash_str`, `marketplace`, `id`, `diversity`}
}
\value{
1. `parsedUsers` dataframe with `profile` column replaced by
  `profileClean`, which has any PGP keys extracted. 2. `retrievedPGPs`
  dataframe with columns `date`, `vendor_hash` and `PGPclean` (extracted PGP
  keys)
}
\description{
Given a parsed profile scrapes, containing columns `marketplace`, `date`,
`id`, `profile`, remove rows with missing information, or duplicated rows,
match to user database and remove non-matches, and extract any available PGP
keys.
}
