% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFit.R
\docType{data}
\name{exampleFit}
\alias{exampleFit}
\title{Example random forest model fit}
\format{A random forest object. For more details see \url{https://github.com/xhtai/heisenbrgr}.}
\usage{
exampleFit
}
\description{
Example random forest model fit
}
\keyword{datasets}
