% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1b_cleanParsedData.R
\name{extractPGPs}
\alias{extractPGPs}
\title{Extract any PGP keys from an input string}
\usage{
extractPGPs(inputString)
}
\arguments{
\item{inputString}{string to search for PGP keys}
}
\value{
1. `outputString` input string removing PGP keys 2. `retrievedPGPs`
  vector of extracted PGP keys (header, footer, whitespace etc. removed).
  Will have more than one element if there is more than 1 unique PGP key in
  the input string
}
\description{
Given an input string, extract any PGP keys (any string starting with
-----BEGIN PGP PUBLIC KEY BLOCK----- and ending with -----END PGP PUBLIC KEY
BLOCK). Note that trailing dashes are not currently removed.
}
