% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_generateComparisons.R
\name{fromList}
\alias{fromList}
\title{Generate pairwise comparisons from list items}
\usage{
fromList(pairsOfInterest, listName, functionName)
}
\arguments{
\item{pairsOfInterest}{dataframe with columns `hash1`, `hash2` referring to
vendor hashes for the comparison}

\item{listName}{e.g. `titleTokens`, `profileTokens`, `PGPlist`}

\item{functionName}{either `jaccardSimilarity` or `PGPmatch`. These are
currently the only two options.}
}
\value{
vector the length of `nrow(pairsOfInterest)`, containing the
  resulting similarity scores for each pair
}
\description{
Given a list item indexed by vendor hashes (such as `titleTokens`, ...,
generated using the functions in `step2_anytime.R`), generate Jaccard
similarities / return whether or not their intersections are non-empty (used
for PGPs).
}
