% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_anyTime.R
\name{generatePGPs}
\alias{generatePGPs}
\title{Get list of PGP keys for each account in some timeframe of interest, from
scrapes of profiles and item descriptions}
\usage{
generatePGPs(PGPclean, userHashesOfInterest, startDate, endDate)
}
\arguments{
\item{PGPclean}{dataframe as described above}

\item{userHashesOfInterest}{vendor_hashes for accounts that have feedback in
timeframe of interest; can be generated by `infoFromFeedback()`}

\item{startDate}{string for start of time frame of interest, e.g. "2014-01-10"}

\item{endDate}{string for end of time frame of interest (exclusive), e.g.
"2014-01-15" means up to 1/14}
}
\value{
`PGPlist` list item
}
\description{
PGP scrapes data are extracted from profile and item scrapes, and needs to
contain the following variables:
\itemize{
    \item{`date`: date that scrape was done, in UNIX time}
    \item{`vendor_hash`: unique identifier for seller account}
    \item{`PGPclean`: extracted PGP key}
}
}
\details{
We first consider all scrapes within the start and end dates specified. If
there are no PGP keys available, we look at equal time periods before and
after the dates of interest. For example, if we are interested in August
2014, and no keys are available then, we consider all scrapes in
July-September 2014.

Returns `PGPlist`, a list of PGP keys for accounts that have feedback in the
timeframe of interest.
}
