% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4b_clustering.R
\name{getClust}
\alias{getClust}
\title{Going from links to clusters}
\usage{
getClust(allPairwise, linkCol, pairColNums)
}
\arguments{
\item{allPairwise}{name of data frame containing all pairwise comparisons.
This needs to have at least four columns, one representing the first item
in the comparison, one representing the second item, and one representing
whether the pair is linked in the given clustering. These are enumerated in
the next two parameters.}

\item{linkCol}{name of column in `allPairwise` with links (a binary indicator
for whether the pair is linked after hierarchical clustering), input as
character, e.g. "minimax0.4"}

\item{pairColNums}{vector of length 2 indicating the column numbers in
`allPairwise` of 1. item 1 in comparison, 2. item 2 in comparison}
}
\value{
data frame with two columns: `item`, the name of the item, and
  `cluster`, the cluster number the item is a member of
}
\description{
Given pairwise links as generated by `linkAnalysis()`, produce a list of
individual items and their cluster membership
}
