% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_generateComparisons.R
\name{getPairs}
\alias{getPairs}
\title{Generate basic pairwise comparisons for user accounts of interest}
\usage{
getPairs(dataframeName)
}
\arguments{
\item{dataframeName}{name of dataframe that contains at least three columns:
user hashes of interest (named `vendor_hash`), alias (`id`) and marketplace
(`marketplace`).}
}
\value{
dataframe with columns `hash1`, `hash2`, `order1` (row corresponding
  to `hash1` in `dataframeName`), `order2`, `id1`, `id2`, `marketplace1`,
  `marketplace2`, `idDist`, `sameMarket`
}
\description{
Given a dataframe containing at least account hashes, IDs and marketplaces,
generate basic pairwise comparisons: edit distance between the IDs and
whether or not they are from the same marketplace.
}
