% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_model.R
\name{hierarchical}
\alias{hierarchical}
\title{Hierarchical clustering to generate final clusters}
\usage{
hierarchical(pairs, final)
}
\arguments{
\item{pairs}{pairs with `preds` column, which are predictions from the random
forest model}

\item{final}{output from Step 2 with all the account-level information}
}
\value{
dataframe of accounts with their final clusters and sizes
}
\description{
Given predictions for each pair, generate transitive closures using
hierarchical clustering with single linkage.
}
