% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_anyTime.R
\name{infoFromDescriptionClean}
\alias{infoFromDescriptionClean}
\title{Get account-level information from item description scrapes for some
timeframe of interest}
\usage{
infoFromDescriptionClean(descriptionClean, userHashesOfInterest,
  itemHashesOfInterest, startDate, endDate)
}
\arguments{
\item{descriptionClean}{dataframe as described above}

\item{userHashesOfInterest}{vendor_hashes for accounts that have feedback in
timeframe of interest; can be generated by `infoFromFeedback()`}

\item{itemHashesOfInterest}{item_hashes for items that have feedback in
timeframe of interest; can be generated by `infoFromFeedback()`}

\item{startDate}{string for start of time frame of interest, e.g. "2014-01-10"}

\item{endDate}{string for end of time frame of interest (exclusive), e.g.
"2014-01-15" means up to 1/14}
}
\value{
list containing 1. dataframe with vendor_hash and
  numDescriptionTokens, 2. `descriptionTokens` list
}
\description{
Item scrapes data needs to contain the following variables:
\itemize{
    \item{`date`: date that scrape was done, in UNIX time}
    \item{`item_hash`: unique identifier for item listed}
    \item{`vendor_hash`: unique identifier for seller account}
    \item{`descriptionClean`: scrape of item description page at that time}
}
}
\details{
We first consider all scrapes within the start and end dates specified. If
there are no scrapes available, we look at equal time periods before and
after the dates of interest. For example, if we are interested in August
2014, and no scrapes are available then, we consider all scrapes in
July-September 2014.

Account-level variables returned are:
\itemize{
    \item{`vendor_hash`: unique identifier for seller account}
    \item{`numDescriptionTokens`: number of unique tokens from item description scrapes}
    \item{`descriptionTokens`: list item}
}
}
\examples{
\dontrun{
    descriptionCleanOut <- infoFromDescriptionClean(descriptionClean,
    out$vendorHashes, out$itemHashes, "2014-08-01", "2014-08-31")
}
}
