% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_anyTime.R
\name{infoFromFeedback}
\alias{infoFromFeedback}
\title{Get account-level information from feedback data for some timeframe of
interest}
\usage{
infoFromFeedback(feedback, startDate, endDate)
}
\arguments{
\item{feedback}{dataframe as described above}

\item{startDate}{string for start of time frame of interest, e.g. "2014-01-10"}

\item{endDate}{string for end of time frame of interest (exclusive), e.g.
"2014-01-15" means up to 1/14}
}
\value{
list containing 1. dataframe with account-level variables as listed
  above, 2. unique vendor_hashes of interest (receiving feedback in period of
  interest), 3. unique item_hashes of interest
}
\description{
Feedback data refers to unique feedback left on the marketplaces. Feedback
needs to contain the following variables:
\itemize{
    \item{`hash_str`: this is a unique identifier for each feedback}
    \item{`date`: date that the feedback was left}
    \item{`vendor_hash`: unique identifier for seller account}
    \item{`item_hash`: unique identifier for item that feedback was left for}
    \item{`marketplace`: marketplace that the feedback was left on}
    \item{`order_amount_usd`: cost of item that feedback was left for}
}
}
\details{
Account-level variables returned are:
\itemize{
    \item{`vendor_hash`: unique identifier for seller account}
    \item{`marketplace`: marketplace that the feedback was left on}
    \item{`totalFeedback`: total number of feedback that the account received}
    \item{`numListingsWithFeedback`: number of item listings where account received feedback}
    \item{`firstFeedback`: date of first feedback (character)}
    \item{`lastFeedback`: date of last feedback (character)}
    \item{`daysActive`: number of days between first and last feedback (inclusive)}
    \item{date columns: binary variables indicating whether feedback was received on these days}
    \item{`dailyFrac`: (totalFeedback/daysActive for vendor) / (marketplaceTotalFeedback/marketplaceDaysActive)}
    \item{`meanPriceSold`}
    \item{`medianPriceSold`}
    \item{`minPriceSold`}
    \item{`maxPriceSold`}
    \item{`priceRange`}

}
}
\examples{
\dontrun{
    out <- infoFromFeedback(feedback, "2014-08-01", "2014-08-31")
}
}
