% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_anyTime.R
\name{infoFromItems}
\alias{infoFromItems}
\title{Get account-level information from items data for some timeframe of interest}
\usage{
infoFromItems(items, itemHashesOfInterest, userHashesOfInterest)
}
\arguments{
\item{items}{dataframe as described above}

\item{itemHashesOfInterest}{item_hashes for items that have feedback in
timeframe of interest; can be generated by `infoFromFeedback()`}

\item{userHashesOfInterest}{vendor_hashes for accounts that have feedback in
timeframe of interest; can be generated by `infoFromFeedback()`}
}
\value{
list containing 1. dataframe with vendor_hash and numTitleTokens, 2.
  `titleTokens` list, 3. `inventory` list
}
\description{
Items data refers to data on unique items listed. Items needs to contain the
following variables:
\itemize{
    \item{`hash_str`: this is a unique identifier for each item listing}
    \item{`marketplace`: marketplace that the item was listed on}
    \item{`title`: title of item listing}
    \item{`vendor`: vendor account ID that listed the item}
    \item{`vendor_hash`: unique identifier for vendor, corresponds to `hash_str` in `users` and feedback}
    \item{`prediction`: category predicted for item listing, see Soska and Christin (2015)}
    \item{`dosage`: dosage extracted from item title}
    \item{`unit`: units extracted from item title}
}
}
\details{
Account-level variables returned are:
\itemize{
    \item{`vendor_hash`: unique identifier for seller account}
    \item{`numTitleTokens`: number of unique tokens from item titles}
    \item{`titleTokens`: list item}
    \item{`inventory`: list item}
}
}
\examples{
\dontrun{
    itemsOut <- infoFromItems(items, out$itemHashes, out$vendorHashes)
}
}
