% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_anyTime.R
\name{infoFromProfileClean}
\alias{infoFromProfileClean}
\title{Get account-level information from profile scrapes for some timeframe of
interest}
\usage{
infoFromProfileClean(profileClean, userHashesOfInterest, startDate,
  endDate)
}
\arguments{
\item{profileClean}{dataframe as described above}

\item{userHashesOfInterest}{vendor_hashes for accounts that have feedback in
timeframe of interest; can be generated by `infoFromFeedback()`}

\item{startDate}{string for start of time frame of interest, e.g. "2014-01-10"}

\item{endDate}{string for end of time frame of interest (exclusive), e.g.
"2014-01-15" means up to 1/14}
}
\value{
list containing 1. dataframe with vendor_hash and numProfileTokens, 2.
  `profileTokens` list
}
\description{
Profile scrapes data needs to contain the following variables:
\itemize{
    \item{`date`: date that scrape was done, in UNIX time}
    \item{`vendor_hash`: unique identifier for seller account}
    \item{`profileClean`: scrape of profile page at that time}
}
}
\details{
We first consider all scrapes within the start and end dates specified. If
there are no scrapes available, we look at equal time periods before and
after the dates of interest. For example, if we are interested in August
2014, and no scrapes are available then, we consider all scrapes in
July-September 2014.

Account-level variables returned are:
\itemize{
    \item{`vendor_hash`: unique identifier for seller account}
    \item{`numProfileTokens`: number of unique tokens from profile scrapes}
    \item{`profileTokens`: list item}
}
}
\examples{
\dontrun{
    profileCleanOut <- infoFromProfileClean(profileClean, out$vendorHashes,
    "2014-08-01", "2014-08-31")
}
}
