% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_anyTime.R
\name{infoFromUsers}
\alias{infoFromUsers}
\title{Takes subset of users data for some timeframe of interest}
\usage{
infoFromUsers(users, userHashesOfInterest)
}
\arguments{
\item{users}{dataframe as described above}

\item{userHashesOfInterest}{vendor_hashes for accounts that have feedback in
timeframe of interest; can be generated by `infoFromFeedback()`}
}
\value{
dataframe with `vendor_hash`, `id`, `diversity`
}
\description{
Users data refers to data on unique user accounts. users
needs to contain the following variables:
\itemize{
    \item{`hash_str`: this is a unique identifier for each user account}
    \item{`marketplace`: marketplace that the feedback was left on}
    \item{`id`: alias used on marketplace}
    \item{`diversity`: diversity coefficient, see Soska and Christin (2015)}
}
}
\examples{
\dontrun{
    usersOut <- infoFromUsers(users, out$vendorHashes)
}
}
