% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1b_cleanParsedData.R
\name{noInfo}
\alias{noInfo}
\title{Check if an input string has useful information}
\usage{
noInfo(inputString)
}
\arguments{
\item{inputString}{string to check}
}
\value{
logical, TRUE if no useful information.
}
\description{
If an input string has only whitespace, "\n" or "-", return TRUE and
otherwise, return FALSE.
}
