% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{runStep1}
\alias{runStep1}
\title{Run all of step 1}
\usage{
runStep1(feedback, items, users, parsedUsers, parsedItems)
}
\arguments{
\item{feedback}{name of dataframe with feedback data}

\item{items}{name of dataframe with unique item listings}

\item{users}{name of dataframe with unique user accounts}

\item{parsedUsers}{name of dataframe with raw profile scrapes}

\item{parsedItems}{name of dataframe with raw item scrapes}
}
\value{
feedback, items, users, profileClean, descriptionClean, PGPclean
}
\description{
Run all of step 1
}
