% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{runStep1_2}
\alias{runStep1_2}
\title{Run all of step 1 (excluding extracting inventory)}
\usage{
runStep1_2(feedback, items, users, parsedUsers, parsedItems)
}
\arguments{
\item{feedback}{name of dataframe with feedback data}

\item{items}{name of dataframe with unique item listings}

\item{users}{name of dataframe with unique user accounts}

\item{parsedUsers}{name of dataframe with raw profile scrapes}

\item{parsedItems}{name of dataframe with raw item scrapes}
}
\value{
feedback, items, users, profileClean, descriptionClean, PGPclean
}
\description{
Run all of step 1 (excluding extracting inventory)
}
