% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{runStep3}
\alias{runStep3}
\title{Run all of step 3 to get pairwise comparisons}
\usage{
runStep3(final, profileTokens, titleTokens, descriptionTokens, inventory,
  PGPlist)
}
\arguments{
\item{final}{dataframe with account-level information}

\item{profileTokens}{list item}

\item{titleTokens}{list item}

\item{descriptionTokens}{list item}

\item{inventory}{list item}

\item{PGPlist}{list item}
}
\value{
allPairwise
}
\description{
Note that this function cannot be used if you do not have the complete set of
account information. If so, step 3 will have to be done manually, removing
lines that there is no information on
}
