% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{runStep2}
\alias{runStep2}
\title{Run all of step 2 to get account-level information}
\usage{
runStep2(feedback, items, users, profileClean, descriptionClean, PGPclean,
  startDate, endDate)
}
\arguments{
\item{feedback}{processed feedback data}

\item{items}{processed items data}

\item{users}{processed users data}

\item{profileClean}{profile information from profile scrapes}

\item{descriptionClean}{item descriptions from item scrapes}

\item{PGPclean}{PGP keys from profile and item scrapes}
}
\value{
final, profileTokens, titleTokens, descriptionTokens, inventory, PGPlist
}
\description{
Run all of step 2 to get account-level information
}
